Hoping this is working, I've added
Progress bar thingies that will be instantiated based on an array saved in userdefaults.
The progress bar thingies have a start date and an end date that they calculate how far is left and that's what's shown in the progress bar, they should update based on how much time is left

I don't have a way of adding more progress bar thingies atm, but that's what I'm gonna work on next, hopefully some ui or something.