I've done all of this work so that I can have a singular variable. That variable is cutRepeatingNotificationTimes
It includes the next 64 future notifications for repeating notifications. I can schedule the now sorta.

Before I go into scheduling notifications I want to compare this list with already existing scheduled notifications. Theres a good chance that this will be run fairly often, I don't want to delete all notifications. I want to schedule notifications that are in cutRepeatingNotificationTimes but not in the list of already scheduled notifications. I want to remove notifications that are in scheduled notifications but not in cutRepeatingNotificationTimes, and I want to do nothing if it's in both.

sorta funky but I think it'll work